pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.125.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

